third_party/blink/web_tests/external/wpt/css/css-borders/tentative/parsing/box-shadow-spread-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['box-shadow-spread'] = "0" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['box-shadow-spread'] = "1px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['box-shadow-spread'] = "1em" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['box-shadow-spread'] = "-1px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['box-shadow-spread'] = "calc(1em + 2px)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
